requirements.txt
requests

django
django-preferences
django-dynamic-preferences
django-modeladmin-reorder
django-settings-export

pytelegrambotapi
